@@ -1,36 +1,32 @@
-import React, { Component } from 'react';
+import React, { useState, useEffect } from 'react';
 import stations from '../stations';
 import Stations from '../Components/Stations/Stations';
 import Player from '../Components/Player/Player';
 import { API_URL } from '../config';
-import { error } from 'util';
 import { Header } from './style';
 
 let interval = null;
+const initialState = {
+	activeStation: {},
+	isPaused: true,
+	trackInfo: {}
+};
 
-export default class App extends Component {
-	constructor(props) {
-		super(props);
-		this.state = { 
-			activeStation: {},
-			isPaused: true,
-			trackInfo: {}
-		};
-	}
+const App = () => {
+	const [state, changeState] = useState(initialState);
+	const { activeStation, isPaused, trackInfo } = state;
 
-	componentDidUpdate() {
-		const { isPaused } = this.state;
+	useEffect(() => {
 		if (!isPaused) {
 			document.getElementById('audio').play();
 		}
-	}
+	}, [isPaused]);
 	
 	// set current playing radiostation to state, for hightlighting ACTIVE station by another color 
-	setActiveRadiostation = station => {
+	const setActiveRadiostation = station => {
 		clearInterval(interval);
-		this.setState({ activeStation: station, isPaused: false });
-		interval = setInterval(this.getTrackInfo, 2000);
-
+		changeState(prevState => ({ ...prevState, activeStation: station, isPaused: false }));
+		interval = setInterval(getTrackInfo, 2000);
 
 		fetch('http://back.com', {
 			method: 'post',
@@ -41,45 +37,42 @@ export default class App extends Component {
 			body: JSON.stringify(station)
 		})
 			.then(res => res.json())
-			.then(res => console.log(res))
 			.catch(err => console.error(err));
-	}
+	};
 
 	// toggle icon on player and set the music play/pause
-	togglePauseIcon = () => {
+	const togglePauseIcon = () => {
 		let player = document.getElementById('audio');
 		if (!player.paused) {
 			player.pause();
-			this.setState({ isPaused: true });
+			changeState(prevState => ({ ...prevState, isPaused: true }));
 		} else {
 			player.play();
-			this.setState({ isPaused: false });
+			changeState(prevState => ({ ...prevState, isPaused: false }));
 		}
-	}
+	};
 
 	// set request to API for track information (author, trackname, download link etc...)
-	getTrackInfo = () => {
-		const { activeStation } = this.state;
+	const getTrackInfo = () => {
 		if (!activeStation.prefix) {
 			return new Error('No active stations!');
 		}
 		const validUrl = `${API_URL}xml/${activeStation.prefix}_online_v8.txt`;
 		fetch(validUrl)
 			.then(res => res.json())
-			.then(data => this.setState({ trackInfo: data }))
+			.then(data => changeState(prevState => ({ ...prevState, trackInfo: data })))
 			.catch(err => console.error(err));
-	}
+	};
 
-	render() {
-		const { activeStation, isPaused, trackInfo } = this.state;
 
-		return (
-			<div className='App'>
-				<Header>
-					<Player activeStation={activeStation} isPaused={isPaused} togglePauseIcon={this.togglePauseIcon} getTrackInfo={this.getTrackInfo} trackInfo={trackInfo} />
-				</Header>
-				<Stations stations={stations} activeStation={activeStation} setActiveRadiostation={this.setActiveRadiostation} />
-			</div>
-		);
-	}
-}
+	return (
+		<div className='App'>
+			<Header>
+				<Player activeStation={activeStation} isPaused={isPaused} togglePauseIcon={togglePauseIcon} getTrackInfo={getTrackInfo} trackInfo={trackInfo} />
+			</Header>
+			<Stations stations={stations} activeStation={activeStation} setActiveRadiostation={setActiveRadiostation} />
+		</div>
+	);
+};
+
+export default App;
