@@ -4,5 +4,5 @@ export const Header = styled.header `
 	display: flex;
     flex-flow: row wrap;
     justify-content: center;
-    background-image: linear-gradient(darkgrey, #383838);
+    background: #c3c3c3;
 `;
